[FFM-5484] Reset cache on init if exception is thrown
What
If we get an exception while loading the initial list of evaluations from the cache then assume cache has gotten corrupt on disk and reset it.

Why
We're seeing type exceptions on the initialize() method on startup which looks to be caused by a corrupt cache on disk. This will remove any bad entries.

Testing
Manual - changed the underlying xml to get it to throw and exception