FFM-6396 - Android/Flutter SDK: Can't initialise SDK without Target Identifier
What
Validate and check target for null or empty identifiers and set to name. If neither name or id are given and set to some default values.
SDK hangs if it gets a 403 back on auth and doesn't report any error in the logs - so check for authentication failures and throw an exception instead of continuing

Why
Passing in a null identifier in the target causes auth endpoint to fail and return 403, which internally sets authInfo to null

Testing
Tested manually with GettingStarted example