FFM-6713 - Android SDK - Evaluation functions not using cache
What
Fix FeatureRepositoryImpl to prefer cache, call out to network to get flag only if its not found in cache and network is available.

Why
The getEvaluation function is performing very poorly and causing excessive API callouts to the ff-server when it should be using the cache.

Testing
Several new unit tests added + manual