FFM-8551 - Android SDK - Insecure TLS/SSL trust manager lint warning on Android
What
Remove OpenAPI code which contains logic to disable SSL by creating empty truststores.
We don't use this code. It is automatically generated by OpenAPI and there doesn't
seem to be any options to disable it, upgrading won't help either so we will
remove it and replace it with an exception if called.

Why
The Gradle link task is flagging this code as insecure with:

"checkClientTrusted is empty, which could cause insecure network traffic due to trusting arbitrary TLS/SSL certificates presented by peers"

Testing
Manual smoke testing