[FFM-8529] - Update SSE implementation (#69)
* [FFM-8529] - Update SSE implementation

What
This removes the old oksse implementation and replaces it with an improved event handler. The SSE code is the same code used in the Java SDK and has been ported over.

Why
It aligns the SDK code with the Java SDK for easier maintenance. It also improves resiliency by adding retrying and backoff.

Testing
Manual