[FFM-10477] - Fix null pointer when calling initialize() (#85)
* [FFM-10477] - Fix null pointer when calling initialize()

What
This is a minimum fix to get rid of a null pointer exception that can happen. First approach was to make cloud final to guarantee it can never be null however getInstance() prevents us from doing that since it controls the instance creation. Second approach was to move this.cloud = cloudFactory.cloud() before init thread creation however this results in a NetworkOnMainThreadException.

For now adding a simple null check but the proper fix is to refactor the code into worker threads and avoid assigning class members from different threads.

Plus:
    - Fix a problem where SDK would not reconnect when started with no network
    - Remove duplicate SSE event log

Why
See seeing a race condition between the init and reschedule when an SDK is started with no network

Testing
Manual